Start SetUp
This is only a start set up to get the builder on the run.
The builder seems to do things pretty fine, however the Windows builds do not have their respective ".LOVE" files in neither the 32 bit as the 64 bit versions, and I wonder why. (#2)
The Linux set up also leaves something to be desired.