Conterted to PNG: Yellow lines
It is very much of a setback that the command line graphics convertion utilities I downloaded are incompatible with 16 color pictures.
This requies me to use GIMP for the conversion, but that is actually taking a bit of the long way around.
I hope there is a quicker way to do this :-/